fix(sync): serialize operations across backends
